+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from './includes';
+export * from './beacons';
 export * from './cache';
 export * from './sayt';
 export * from './search';