+++ src/sayt/products.ts
@@ -1,3 +1,4 @@
+import { SendableOrigin } from 'gb-tracker-client/models';
 import { Request, Results } from 'groupby-api';
 import { ErrorPayload, Product, WithQuery, WithGroup } from '../includes';
 
@@ -7,6 +8,8 @@ export const SAYT_PRODUCTS_REQUEST = 'gbe::sayt_products_request';
 export interface SaytProductsRequestPayload extends WithQuery, WithGroup {
   /** The extra configuration options to customize the request. */
   config?: Partial<Request>;
+  /** The origin of the products request action. */
+  origin: keyof SendableOrigin;
 }
 
 /** The name of the event fired when the results of a SAYT products request have been received. */